+++ metalnap/__init__.py
@@ -2,4 +2,4 @@ from .config import Config          # noqa: F401
 from .controller import Controller  # noqa: F401
 from .types import NodeState        # noqa: F401
 
-__version__ = "0.2.0"
+__version__ = "0.2.1"